% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{position_sf_dodge}
\alias{position_sf_dodge}
\title{Horizontally dodge overlapping sf point geometries by group}
\usage{
position_sf_dodge(width = 0.01, bound = NULL)
}
\arguments{
\item{width}{numeric, amount of horizontal spacing between points. Defaults to 0.01.}

\item{bound}{sf polygon object that defines the boundary within which points should stay.
If NULL, points are not constrained.}
}
\value{
A Position object that can be used with geom_sf.
}
\description{
Creates a Position object that arranges overlapping points in a horizontal line
with equal spacing between them, using ggplot2's group aesthetic.
}
