% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{position_sf_triangle}
\alias{position_sf_triangle}
\title{Arrange sf point geometries in a triangular pattern by group}
\usage{
position_sf_triangle(spacing = 0.01, bound = NULL)
}
\arguments{
\item{spacing}{numeric, spacing between points. Defaults to 0.01.}

\item{bound}{sf polygon object that defines the boundary within which points should stay.
If NULL, points are not constrained.}
}
\value{
A Position object that can be used with geom_sf.
}
\description{
Creates a Position object that arranges overlapping points in a triangular pattern,
using ggplot2's group aesthetic.
}
