% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{position_sf_dodge}
\alias{position_sf_dodge}
\title{Horizontally dodge overlapping sf point geometries}
\usage{
position_sf_dodge(width = 0.01, bound = NULL, hjust = 0.5, vjust = 0.5)
}
\arguments{
\item{width}{numeric, amount of horizontal spacing between points. Defaults to 0.01.}

\item{bound}{sf polygon object that defines the boundary within which points should stay.}

\item{hjust}{numeric between 0 and 1, horizontal justification. 0 = left, 1 = right. Defaults to 0.5.}

\item{vjust}{numeric between 0 and 1, vertical justification. 0 = bottom, 1 = top. Defaults to 0.5.}
}
\description{
Horizontally dodge overlapping sf point geometries
}
